Refactor directory structure to enable 'from msgraph.core' import format
