+++ msgraph/__init__.py
@@ -1,0 +1,4 @@
+# ------------------------------------
+# Copyright (c) Microsoft Corporation.
+# Licensed under the MIT License.
+# -----------------------------------
