+++ msgraph/core/__init__.py
@@ -1,0 +1,5 @@
+from .client_factory import HTTPClientFactory
+from .constants import SDK_VERSION
+from .graph_client import GraphClient
+
+__version__ = SDK_VERSION
