+++ msgraph/core/constants.py
