+++ msgraph/core/enums.py
