+++ msgraph/core/middleware/middleware.py
@@ -4,7 +4,7 @@ import uuid
 from requests.adapters import HTTPAdapter
 from urllib3 import PoolManager
 
-from msgraphcore.middleware.request_context import RequestContext
+from .request_context import RequestContext
 
 
 class MiddlewarePipeline(HTTPAdapter):
