--- msgraphcore/__init__.py
@@ -1,5 +1,0 @@
-from msgraphcore.client_factory import HTTPClientFactory
-from msgraphcore.constants import SDK_VERSION
-from msgraphcore.graph_client import GraphClient
-
-__version__ = SDK_VERSION
