+++ tests/unit/test_client_factory.py
@@ -2,11 +2,11 @@ import pytest
 from requests import Session
 from requests.adapters import HTTPAdapter
 
-from msgraphcore.client_factory import HTTPClientFactory
-from msgraphcore.constants import CONNECTION_TIMEOUT, REQUEST_TIMEOUT
-from msgraphcore.enums import APIVersion, NationalClouds
-from msgraphcore.middleware.authorization import AuthorizationHandler
-from msgraphcore.middleware.middleware import BaseMiddleware, MiddlewarePipeline
+from msgraph.core.client_factory import HTTPClientFactory
+from msgraph.core.constants import CONNECTION_TIMEOUT, REQUEST_TIMEOUT
+from msgraph.core.enums import APIVersion, NationalClouds
+from msgraph.core.middleware.authorization import AuthorizationHandler
+from msgraph.core.middleware.middleware import BaseMiddleware, MiddlewarePipeline
 
 
 def test_initialize_with_default_config():
