+++ tests/unit/test_middleware_pipeline.py
@@ -1,7 +1,7 @@
 from collections import OrderedDict
 from unittest import TestCase
 
-from msgraphcore.middleware.middleware import BaseMiddleware, MiddlewarePipeline
+from msgraph.core.middleware.middleware import BaseMiddleware, MiddlewarePipeline
 
 
 class MiddlewarePipelineTest(TestCase):
